requirements.txt
astroid==1.5.3
backports.functools-lru-cache==1.4
configparser==3.5.0
Django==1.11.5
djangorestframework==3.6.4
djangorestframework-jwt==1.8.0
enum34==1.1.6
gunicorn==19.7.1
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
PyJWT==1.4.2
pylint==1.7.2
pytz==2017.2
singledispatch==3.4.0.3
six==1.10.0
wrapt==1.10.11
